+++ packages/docs/site/rollup.config.js
@@ -38,11 +38,6 @@ export default {
 				emitCss: true,
 			}),
 			resolve({
-
-				/* disabled as this makes the build fail with
-				> 'makeStyle' is not exported by ../../tools/dom/dist/browser.js, imported by ../../components/barchart/src/BarchartV.svelte
-				*/
-				// browser: true,
 				dedupe: ['svelte']
 			}),
 			commonjs(),