+++ packages/tools/geo/src/modules/topojson.js
@@ -79,10 +79,10 @@ export const topoToGeo = (topojson, id) =>
 > topoRO = filterGeometriesBy('RO')
 >
 > topoRO(NUTS_RG_03M_2003_4326_LEVL_0)
-// see `data/dist/topojson/NUTS_RG_03M_2003_4326_LEVL_0_RO.json`
+// see `svizzle/atlas/data/dist/topojson/NUTS_RG_03M_2003_4326_LEVL_0_RO.json`
 >
 > topoRO(NUTS_RG_03M_2006_4326_LEVL_0)
-// see `data/dist/topojson/NUTS_RG_03M_2006_4326_LEVL_0_RO.json`
+// see `svizzle/atlas/data/dist/topojson/NUTS_RG_03M_2006_4326_LEVL_0_RO.json`
  *
  * @since 0.8.0
  */