+++ packages/components/ui/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - moved `@macfja/svelte-persistent-store` to dependencies to fix the `StorageIO` component
 - added `FetchDriver` component wrapping access to `@svizzle/request/src/fetchManager/fetchManager.js`
+- line height in A11y menu is now unitless
+- added A11y menu components to `ui/src/index.js`
 
 ## `@svizzle/ui` v0.5.0